Add keybindings and refactor a bit
